refactor styles for filtersbar
